requirements.txt
# with python3.12
flask==3.0.3
flask_sqlalchemy==3.1.1
requests==2.32.2
python-dotenv==1.0.1
gunicorn==20.0.4
werkzeug==3.0.3